Update examples/pubsub-simple to show how to retry
The pubsub-simple example now tells dapr to retry delivery on the first
message the subscriber receives.

The message is logged to stdout by dapr and we capture that with mm.

Since we are only retrying the first message, stdout will show the old
output, plus the first line repeated, and the dapr log message.

Signed-off-by: Renato Luiz de Freitas Cunha <renato.cunha@microsoft.com>