Metadata API support for Python GRPC clients. (#446)
This PR adds support for [Metadata API] to the Python SDK,
along with tests and a sample example code.

Two new methods were added to the python GRPC client:

1. **get_metadata**: Gets the Dapr sidecar information provided by
   the Metadata Endpoint.
2. **set_metadata**: Adds a custom label to the Dapr sidecar
   information stored by the Metadata endpoint.

It might be worth pointing out that while this PR is code-complete,
the current GRPC implementation of the [Metadata API] in dapr 1.18
returns a response missing some fields. See dapr/dapr#5038 and
dapr/dapr#5018. This is being tracked in #445.

Fixes #410

[Metadata API]: https://docs.dapr.io/reference/api/metadata_api/

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>
Co-authored-by: Bernd Verst <4535280+berndverst@users.noreply.github.com>