Define SDK dependency on typing-extensions.
PR #521 added an implicit dependency on `typing-extensions`. During development this dependency is satisfied as it is mentioned on `dev-requirements.txt`. However, new instalations of the Python SDK will not bring this pendency and applications depending on `dapr.client.DaprClient` will fail to load with an import exception.

This PR addresses this by adding `typing-extensions` to the dependency list of the generated python sdk package.

Closes #544

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>