Add DaprInternalError.as_json_safe_dict for actors (#765)
The FastAPI and Flask extensions for actors serialise the value of
any raised DaprInternalError to JSON, which fails if the error
contains bytes in its `_raw_response_bytes` field.

This change adds a new `as_json_safe_dict` method and uses it in
place of the `as_dict` method in the FastAPI and Flask extensions.

Two unit tests for the `as_json_safe_dict` method are included.

Signed-off-by: Billy Brown <druidofluhn@gmail.com>
Co-authored-by: Elena Kolevska <elena-kolevska@users.noreply.github.com>